refactor: merge behavior of root branch and branch controller
