--- Assets/_Game/Entities/Creeper/_Branches/BranchController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 59897e50bef1ebf428675700cf88f6c9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
